Former-commit-id: a8bb6c9e68eed9de1efcef6bf0a2e9e2d6008d0b

+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/exercise/request/CreateExerciseRequest.java
@@ -10,13 +10,13 @@ public class CreateExerciseRequest {
 
     private String name; // 운동 명
 
-    private Long setTime; // 근력 : 세트 수, 유산소 : 분
+    private int setTime; // 근력 : 세트 수, 유산소 : 분
 
-    private Long number; // 횟수
+    private int number; // 횟수
 
     private boolean success; // 운동 성공 여부
 
-    private Long weight; //중량
+    private int weight; //중량
 
     private boolean type; // false 근력, true 유산소
