+++ BACK/muscle_maker/src/main/java/com/ssafy/muscle_maker/dto/feeds/request/TagDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class TagDto {
 
-    Long tagId;
+    int tagId;
     String tag;
 
     public Tag toEntity(TagDto tagDto){
